Refactor autoplay feature to separate files
