+++ css/autoplay.css
@@ -1,0 +1,4 @@
+.disabledAutoPlayImg {
+	border-top-left-radius: 1.3em;
+	border-top-right-radius: 1.3em;
+}
