+++ css/browser.css
@@ -316,8 +316,3 @@ body.show-message-buttons ._4_j4 ._39bj {
 	position: relative;
 	left: 100px;
 }
-
-.disabledAutoPlay {
-	border-top-left-radius: 1.3em;
-	border-top-right-radius: 1.3em;
-}
