+++ source/browser/selectors.ts
@@ -1,3 +1,4 @@
 export default {
-	conversationList: 'div[role="navigation"] > div > ul'
+	conversationList: 'div[role="navigation"] > div > ul',
+	conversationSelector: '._4u-c._1wfr > ._5f0v.uiScrollableArea',
 };
